Put list utility functions in separate file
